requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
decouple==0.0.7
Django==5.2.4
django-rest-framework==0.1.0
djangorestframework==3.16.0
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
sqlparse==0.5.3
